requirements.txt
# Core chess analysis libraries
python-chess==1.999
stockfish==3.28.0

# Data analysis and manipulation
pandas==2.1.4
numpy==1.24.3

# Visualization libraries
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Jupyter notebook support
jupyter==1.0.0
ipywidgets==8.1.1
notebook==7.0.6

# API and web requests
requests==2.31.0
urllib3==2.1.0

# Data storage and serialization
jsonlines==4.0.0

# Progress bars and utilities
tqdm==4.66.1

# Configuration management
python-dotenv==1.0.0

# Testing framework
pytest==7.4.3
pytest-cov==4.1.0

# Code formatting and linting
black==23.12.0
flake8==6.1.0